refreshed logic
