+++ cmd/admin-auth-api/main.go
@@ -11,6 +11,7 @@ import (
 	"github.com/ua-academy-projects/share-bite/internal/config"
 	"github.com/ua-academy-projects/share-bite/pkg/closer"
 	"github.com/ua-academy-projects/share-bite/pkg/database/pg"
+	"github.com/ua-academy-projects/share-bite/pkg/jwt"
 	"github.com/ua-academy-projects/share-bite/pkg/logger"
 	"go.uber.org/zap"
 )
@@ -19,11 +20,6 @@ func main() {
 	ctx := context.Background()
 
 	// for local development only
-	/*	if err := config.Load(".env"); err != nil {
-		logger.Fatal(ctx, "load config:", err)
-	}*/
-
-	// docker variant
 	if err := config.Load(".env"); err != nil {
 		logger.Fatal(ctx, "load config:", err)
 	}
@@ -57,7 +53,15 @@ func main() {
 	userRepo := userrepo.New(client)
 
 	// services
-	authSvc := authsvc.New(userRepo)
+	jwtCfg := config.Config().JwtToken
+	jwtManager := jwt.NewTokenManager(
+		jwtCfg.AccessTokenSecretKey(),
+		jwtCfg.RefreshTokenSecretKey(),
+		jwtCfg.AccessTokenTTL(),
+		jwtCfg.RefreshTokenTTL(),
+	)
+
+	authSvc := authsvc.New(userRepo, jwtManager)
 
 	// handlers
 	auth.RegisterHandlers(router.Group("/auth"), authSvc)
