+++ internal/admin-auth/entity/entity.go
@@ -1,0 +1,18 @@
+package entity
+
+type User struct {
+	ID           string
+	Email        string
+	PasswordHash string
+}
+
+type Role struct {
+	ID   int
+	Slug string
+	Name string
+}
+
+type Tokens struct {
+	AccessToken  string
+	RefreshToken string
+}
