+++ internal/admin-auth/handler/auth/handler.go
@@ -2,7 +2,6 @@ package auth
 
 import (
 	"context"
-	"net/http"
 
 	"github.com/gin-gonic/gin"
 	authsvc "github.com/ua-academy-projects/share-bite/internal/admin-auth/service/auth"
@@ -23,81 +22,4 @@ func RegisterHandlers(r *gin.RouterGroup, service authService) {
 	r.POST("/login", h.login)
 	r.POST("/register", h.register)
 	r.POST("/refresh", h.refresh)
-}
-
-type loginRequest struct {
-	Email    string `json:"email"    binding:"required,email"`
-	Password string `json:"password" binding:"required"`
-}
-
-type registerRequest struct {
-	Email    string `json:"email" binding:"required,email,max=254"`
-	Password string `json:"password" binding:"required,min=8,max=72"`
-	Slug     string `json:"slug" binding:"required,oneof=user business"`
-}
-
-func (h *handler) login(c *gin.Context) {
-	var req loginRequest
-	if err := c.ShouldBindJSON(&req); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"message": err.Error()})
-		return
-	}
-
-	tokens, err := h.service.Login(c.Request.Context(), req.Email, req.Password)
-	if err != nil {
-		_ = c.Error(err)
-		return
-	}
-
-	c.JSON(http.StatusOK, gin.H{
-		"access_token":  tokens.AccessToken,
-		"refresh_token": tokens.RefreshToken,
-	})
-}
-
-func (h *handler) register(c *gin.Context) {
-	var req registerRequest
-	if err := c.ShouldBindJSON(&req); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"message": err.Error()})
-		return
-	}
-
-	tokens, err := h.service.Register(
-		c.Request.Context(),
-		req.Email,
-		req.Password,
-		req.Slug,
-	)
-	if err != nil {
-		_ = c.Error(err)
-		return
-	}
-
-	c.JSON(http.StatusCreated, gin.H{
-		"access_token":  tokens.AccessToken,
-		"refresh_token": tokens.RefreshToken,
-	})
-}
-
-type refreshRequest struct {
-	RefreshToken string `json:"refresh_token" binding:"required"`
-}
-
-func (h *handler) refresh(c *gin.Context) {
-	var req refreshRequest
-	if err := c.ShouldBindJSON(&req); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"message": err.Error()})
-		return
-	}
-
-	tokens, err := h.service.Refresh(c.Request.Context(), req.RefreshToken)
-	if err != nil {
-		_ = c.Error(err)
-		return
-	}
-
-	c.JSON(http.StatusOK, gin.H{
-		"access_token":  tokens.AccessToken,
-		"refresh_token": tokens.RefreshToken,
-	})
 }
